Use test directory structure in tracpro.profiles
